requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.17.0
h5py==3.11.0
gunicorn==22.0.0
